compilers: Put clang mixin in a module
