+++ mesonbuild/compilers/objc.py
@@ -17,9 +17,10 @@ import typing
 
 from ..mesonlib import EnvironmentException, MachineChoice
 
-from .compilers import Compiler, ClangCompiler
+from .compilers import Compiler
 from .mixins.clike import CLikeCompiler
 from .mixins.gnu import GnuCompiler
+from .mixins.clang import ClangCompiler
 
 class ObjCCompiler(CLikeCompiler, Compiler):
     def __init__(self, exelist, version, for_machine: MachineChoice, is_cross: bool, exe_wrap: typing.Optional[str]):
